resto-menuservice/requirements.txt
django
djangorestframework
drf-nested-routers
django-cors-headers
psycopg2-binary
gunicorn
requests
